+++ src/electron/views/renderer-initializer.ts
@@ -50,7 +50,7 @@ import { WindowStateStore } from 'electron/flux/store/window-state-store';
 import { createScanResultsFetcher } from 'electron/platform/android/fetch-scan-results';
 import { ScanController } from 'electron/platform/android/scan-controller';
 import { createDefaultBuilder } from 'electron/platform/android/unified-result-builder';
-import { UnifiedSettingsProvider } from 'electron/settings/android-settings-provider';
+import { UnifiedSettingsProvider } from 'electron/settings/unified-settings-provider';
 import {
     RootContainerDeps,
     RootContainerState,
